Relocate disable warning logic to confirmation modal
