+++ web/common/Confirmation.js
@@ -6,24 +6,71 @@ import IconButton from "@material-ui/core/IconButton";
 import { LoadingButton } from "common/components/LoadingButton";
 import { CustomDialogActions, CustomDialogTitle } from "./CustomDialogTitle";
 import DialogContent from "@material-ui/core/DialogContent";
+import Typography from "@material-ui/core/Typography";
+import FormControlLabel from "@material-ui/core/FormControlLabel/FormControlLabel";
+import Checkbox from "@material-ui/core/Checkbox";
+import { useApi } from "common/utils/api";
+import { DISABLE_WARNING_MUTATION } from "common/utils/apiQueries";
 
 export default function ConfirmationModal({
   title,
   textButtons,
   confirmButtonLabel,
   cancelButtonLabel,
+  disableWarningName = null,
   content = null,
   open,
   handleClose,
   handleConfirm
 }) {
+  const api = useApi();
+  const [shouldDisableWarning, setShouldDisableWarning] = React.useState(false);
+
+  React.useEffect(() => {
+    if (open) setShouldDisableWarning(false);
+  }, [open]);
+
+  async function handleConfirmWithEventualWarningDisable(...args) {
+    try {
+      if (disableWarningName && shouldDisableWarning) {
+        await api.graphQlMutate(
+          DISABLE_WARNING_MUTATION,
+          { warningName: disableWarningName },
+          { context: { nonPublicApi: true } }
+        );
+      }
+    } finally {
+      await handleConfirm(...args);
+    }
+  }
+
   return (
     <Dialog onClose={handleClose} open={open}>
       <CustomDialogTitle
         title={title || "Confirmer"}
         handleClose={handleClose}
       />
-      {content && <DialogContent>{content}</DialogContent>}
+      {(content || disableWarningName) && (
+        <DialogContent>
+          {content}
+          {disableWarningName && (
+            <FormControlLabel
+              control={
+                <Checkbox
+                  checked={shouldDisableWarning}
+                  onChange={e => setShouldDisableWarning(e.target.checked)}
+                  size="small"
+                />
+              }
+              label={
+                <Typography variant="caption">
+                  Ne plus afficher ce message
+                </Typography>
+              }
+            />
+          )}
+        </DialogContent>
+      )}
       <CustomDialogActions>
         {cancelButtonLabel || textButtons ? (
           <LoadingButton
@@ -44,7 +91,7 @@ export default function ConfirmationModal({
             color="primary"
             variant="contained"
             onClick={async (...args) => {
-              await handleConfirm(...args);
+              await handleConfirmWithEventualWarningDisable(...args);
               handleClose();
             }}
           >
